+++ src/app/aip/page.tsx
@@ -6,9 +6,9 @@ import { useRouter } from 'next/navigation';
 import axios from 'axios';
 import Image from 'next/image';
 import { format } from 'date-fns';
-import no_record from '../../assets/images/no_record.png';
+import no_record from '@/assets/images/no_record.png';
 import { css } from '@emotion/react';
-import Pagination from '../../src/components/Pagination';
+import Pagination from '@/components/Pagination';
 import { useTranslation } from 'next-i18next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { GetStaticProps } from 'next';
@@ -238,10 +238,10 @@ const strategyListStyle = css`
   }
 `;
 
-export const getStaticProps: GetStaticProps = async ({ locale }) => ({
-  props: {
-    ...(await serverSideTranslations(locale!, ['common'])),
-  },
-});
+// export const getStaticProps: GetStaticProps = async ({ locale }) => ({
+//   props: {
+//     ...(await serverSideTranslations(locale!, ['common'])),
+//   },
+// });
 
 export default StrayegyList;
