+++ src/app/resetPassword/page.tsx
@@ -4,7 +4,7 @@ import React, { useState, useEffect, useRef } from 'react';
 import { useRouter } from 'next/router';
 import '../styles/resetPassword.css';
 import axios from 'axios';
-import {getInvalidFields} from '../utils/validator'
+import { getInvalidFields } from '@/utils/validator';
 
 interface InvalidFields {
   password?: { message: string }[];
@@ -18,16 +18,14 @@ const resetPassword = () => {
   });
   const rules = () => ({
     password: [
-      { required: true, message: ('validator.password_required') },
+      { required: true, message: 'validator.password_required' },
       { validator: validatePass, trigger: 'blur' },
     ],
-    passwordCheck: [
-      { validator: validatePassCheck, trigger: 'blur' },
-    ],
+    passwordCheck: [{ validator: validatePassCheck, trigger: 'blur' }],
   });
   const [invalidFields, setInvalidFields] = useState<InvalidFields>({});
   const [isProccessing, setIsProccessing] = useState(false);
-  
+
   const router = useRouter();
 
   const formRef = useRef<any>(null);
@@ -61,7 +59,7 @@ const resetPassword = () => {
 
   const handleResetPassword = async () => {
     const { token } = router.query;
-    const invalidFields = await getInvalidFields(formData, rules())
+    const invalidFields = await getInvalidFields(formData, rules());
     if (invalidFields) {
       setInvalidFields(invalidFields);
       return;
@@ -70,7 +68,7 @@ const resetPassword = () => {
     try {
       let response = await axios.post('/pub/auth/resetPassword', {
         token,
-        password: formData.password
+        password: formData.password,
       });
       setIsProccessing(false);
       alert('prompt.password_is_reset');
@@ -83,43 +81,60 @@ const resetPassword = () => {
 
   return (
     <div>
-      <section className='section'>
-        <div className='container'>
-          <div className='box'>
-            <div className='header'>
-              <p className='is-size-6 is-pulled-left margin-right: 10px;'>{('caption.reset_password')}</p>
+      <section className="section">
+        <div className="container">
+          <div className="box">
+            <div className="header">
+              <p className="is-size-6 is-pulled-left margin-right: 10px;">
+                {'caption.reset_password'}
+              </p>
             </div>
-            <div className='field'>
-              <label className='label'><span className='has-text-danger'>*</span>{('label.input')}{('label.new_password')}</label>
-              <div className='control has-icons-left'>
-                <input 
-                  className='input' type='password' value={formData.password} placeholder='password'
+            <div className="field">
+              <label className="label">
+                <span className="has-text-danger">*</span>
+                {'label.input'}
+                {'label.new_password'}
+              </label>
+              <div className="control has-icons-left">
+                <input
+                  className="input"
+                  type="password"
+                  value={formData.password}
+                  placeholder="password"
                 />
-                <span className='icon is-small is-left'>
-                    <i className='fa fa-lock'></i>
-                  </span>
+                <span className="icon is-small is-left">
+                  <i className="fa fa-lock"></i>
+                </span>
               </div>
-              {invalidFields.password && <p className='help is-danger' ></p>}
+              {invalidFields.password && <p className="help is-danger"></p>}
             </div>
-            <div className='field'>
-              <label className='label'><span className='has-text-danger'>*</span>{('label.input')}{('label.confirm_password')}</label>
-              <div className='control has-icons-left'>
-                <input 
-                  className='input' type='password' value={formData.passwordCheck} placeholder='Confirm password'
+            <div className="field">
+              <label className="label">
+                <span className="has-text-danger">*</span>
+                {'label.input'}
+                {'label.confirm_password'}
+              </label>
+              <div className="control has-icons-left">
+                <input
+                  className="input"
+                  type="password"
+                  value={formData.passwordCheck}
+                  placeholder="Confirm password"
                 />
-                <span className='icon is-small is-left'>
-                    <i className='fa fa-lock'></i>
-                  </span>
+                <span className="icon is-small is-left">
+                  <i className="fa fa-lock"></i>
+                </span>
               </div>
-              {invalidFields.passwordCheck && <p className='help is-danger' ></p>}
+              {invalidFields.passwordCheck && <p className="help is-danger"></p>}
             </div>
-            <div className='field is-grouped'>
-              <div className='control'>
-                <button 
+            <div className="field is-grouped">
+              <div className="control">
+                <button
                   className={`button is-link is-fullwidth is-focused ${isProccessing ? 'is-loading' : ''}`}
                   onClick={handleResetPassword}
-                  disabled={isProccessing}>
-                  {('action.confirm')}
+                  disabled={isProccessing}
+                >
+                  {'action.confirm'}
                 </button>
               </div>
             </div>
